updated switching between example and tests
