+++ apps/index.ts
@@ -1,4 +1,3 @@
 import App from './App';
-import TestApp from './TestApp';
 
 export default App;
